build: Start the simulator and tpm2-abrmd for extended-sessions test.
This machinery was never added to this test since it was disabled.
Now that's it's been reenabled and added to XFAIL the failure will
still happen but we can make the failure meaningful by integrating
the machinery that starts up the tpm_server and tpm2-abrmd. This will
cause the tests logfile to contain the real reason for the failure
instead of just the output from the TCTI complaining about not being
able to connect to the tpm2-abrmd.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>