WIP: Move session load from tpm2_session_save to tpm2_session_restore.
When a tool wants the session its using to persist across tool
invocation then it must save the context and save this to disk so that
the next tool that comes along can reload it using ContextLoad.
Previously the tools `tpm2_session_save` function was saving and then
immediately loading the session. The handle was passed to the next tool
but the session had already been flushed.

So the session save logic has been changed to save the session context,
write the tpm2_session structure to disk and we now append the context
to this same file using `files_save_tpm_context_to_file`. In
tpm2_session_restore we read the tpm2_session structure back, then read
the context and load it immediately.

This has implications with regard to where the code that load the
session can be executed since we need access to the TPM2_SYS_CONTEXT
when loading the session context. In each tool we'll have to move all
calls to `tpm2_auth_util_from_optarg` from the `on_option` callback to
the `tpm2_tool_onrun` function.