tpm2_eventlog: Add tool to parse the TPM2 event log.
The format for the event log parsed by this tool is documented in
section 9 of the "TCG PC Client Platform Firmware Profile
Specification" (aka "PFP"). This is the "crypto-agile" format capable of
representing digests of varying size. Some constants / values necessary
for understanding the log contents are spread out over other specs but
the structure is documented in the PFP.

This log is a series of concatinated binary structures describing events
generated / collected by the firmware on platform boot. Each 'event' has
3 primary components: First is a header describing the PCR extended, the
'type' of the event and the number of digests in the event. After this a
series of digest structures follows. The final data in this structure is
the event size and a binary representation of the event. The total size
of the log must be inferred.

This tool takes the log file as it's only parameter (positional). It
parses the log and writes an ascii / text representation to stdout. The
option processing and 'main' function for the tool is located in
'tools/misc/tpm2_eventlog.c'. If the log file parameter is omitted the
tool will fallback to parsing the event log exposed by the kernel via
securityfs (/sys/kernel/security/tpm0/binary_bios_measurements).

The log parsing logic and the text formatting logic are located in the
'lib/tpm2_eventlog' and 'lib/tpm2_eventlog_ascii' modules respectively.
The main driver initiates the log parsing through the 'foreach_event2'
function that takes a reference to the start of the log, its size and a
callback function. This callback is invoked by the parsing logic for
each event structure encountered in the log. It's passed a reference to
the event and the size of the event (validated by the parser before
invoking the callback). The display logic is implemented as callbacks
invoked the the parser.

Additionally this commit adds unit tests and a single integration test.
Unit tests are used to test the parser and display logic. The single
integration test invokes the tool passing it a single binary event
structure taken from a VM booting the OVMF firmware with an emulated
TPM2 backend.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>